% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_count_json.R
\name{EvCountjson}
\alias{EvCountjson}
\title{EvCountjson}
\usage{
EvCountjson(KORA.Photo.Output, Event_length, species)
}
\arguments{
\item{Event_length}{Time window to be considered as independent in minutes (ex. 30)}

\item{KORA.Photo.Output:}{Name folder (.json)}

\item{species:}{Species name in latin (ie "Lynx lynx")}
}
\value{

}
\description{
This function creates a table with the number of event for a given species.
}
