% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvCountR.R
\name{EvCountR}
\alias{EvCountR}
\title{EvCountR}
\usage{
EvCountR(KORA.Photo.Output, Event_length, species)
}
\arguments{
\item{KORA.Photo.Output}{Name of the object in R consol. Should be the KORA ouput with lynx or any other species observed (All have been IDed)}

\item{species}{}
}
\value{

}
\description{
This function creates a table with the number of event for a given species.
}
