% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multimark_functions.R
\name{MltiM.Closed.Sometimes.TM}
\alias{MltiM.Closed.Sometimes.TM}
\title{Multimark Model Closed Sometimes}
\usage{
MltiM.Closed.Sometimes.TM(
  KORA.Photo.Output,
  Lynx.Master.Output,
  Period.day,
  simultaneous.minute
)
}
\arguments{
\item{#}{Author: Le Grand}
}
\value{

}
\description{
This function creates a table to be used for the model Multimark Closed Sometimes
}
