% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test.function}
\alias{test.function}
\title{Test Function}
\usage{
test.function(KORA.Photo.Output, Event_length, species)
}
\arguments{
\item{#}{Author: Luc Le Grand}
}
\value{

}
\description{
Test Function
}
