% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_count_json.R
\name{EvCountjson}
\alias{EvCountjson}
\title{EvCountjson}
\usage{
EvCountjson(KORA.Photo.Output, Event_length, species)
}
\arguments{
\item{#}{Author: Luc Le Grand}
}
\value{

}
\description{
EvCountjson
}
