% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_count_R.R
\name{EvCountR}
\alias{EvCountR}
\title{Event Counting}
\usage{
EvCountR(Table_Object, Event_length)
}
\arguments{
\item{#}{Author: Luc Le Grand}
}
\value{

}
\description{
This function creates a table with the number of event for a given species.
Input needed:
   - KORA.Photo.Output = "Nameofcsv.csv" KORA ouput with lynx or any other species observed (All have been IDed)
   Important Note:
   - The csv must be in the folder of the used RStudio Project
   - Date in the csv must be in '%Y-%m-%d %H:%M:%S' (Standard KORA data and R format)
}
