% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multimark_functions.R
\name{MltiM.Closed.Sometimes.TM}
\alias{MltiM.Closed.Sometimes.TM}
\title{Multimark Model Closed Sometimes}
\usage{
MltiM.Closed.Sometimes.TM(
  KORA.Photo.Output,
  Lynx.Master.Output,
  Period.day,
  simultaneous.minute
)
}
\arguments{
\item{infile}{Table Maker}
}
\value{
csv table
}
\description{
Multimark Model Closed Sometimes
    This function can be used to create the Matrix needed for the Multimark model.The following input are needed:
    - KORA.Photo.Output: "Nameofthecsv.csv" ouput from KORA photo (only lynx of the session)
    - Lynx.Master.Output: "Nameofthecsv.csv" output from KORA data
    - Period.day: How long should be the period (ex. Period.day=3)
    - What is considered as simultaneous sighting? (ex. simultaneous.minute=30)

    !!! Important Note:
    - CSV should be in the folder of your RStudio Project
    - date should be in R standard ("%Y-%m-%d %H:%M:%S")

}
